admin_dashboard/requirements.txt
Django>=5.2.8
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2